==> analysis/R/3_exploratory_plots_maps.Rmd <==
---
title: Exploratory plots and maps
author: Guy Lomax
date: 2024-05-14
output: html_notebook
---

Generates plots and maps of GPP, land cover and covariates used in the study region.

```{r setup, include=FALSE}

# If running for first time, run renv::restore() to install required packages
# renv::restore()

# Data management

library(tidyverse)
library(terra)
library(sf)
library(here)

# Visualisation
library(tmap)
library(gratia)
library(ggpubr)

```


Load data

```{r load_data, include = FALSE}

# Africa country boundaries (Natural Earth dataset)
countries <- st_read(here("data", "raw", "vector", "natural_earth",
                          "ne_110m_admin_0_countries_fixed.shp"))
africa <- filter(countries, CONTINENT == "Africa")
africa_vct <- vect(africa)

# Load multi-annual data frame for modelling

df <- read_csv(here("data", "processed", "csv", "df_annual.csv")) %>%
  select(cell, lc, x, y, gpp, starts_with("ppt"), starts_with("season"), everything())

# Add land cover class names
igbp_index <- tibble(
  lc = c(7, 9, 10),
  cover = factor(c("Open shrublands",
                   "Open savannas",
                   "Grasslands")))
df_cover <- inner_join(df, igbp_index)

# Load variable rasters for maps
var_stack_all <- rast(here("data", "processed", "raster", "var_raster_stack.tif"))
study_area_mask <- rast(here("data", "processed", "raster", "study_area_mask.tif"))

```

Generate covariate scatter plots

```{r scatter_plots, eval = FALSE, echo = FALSE}

plot_dir <- here("results", "figures")

# Correct precipitation anomaly definition
df_cover_corrected <- mutate(df_cover, pptAnomaly = (pptAnomaly - 1) * 100)


# Two-variable plots
# Sample to reduce display time and overplotting
set.seed(123)
plot_sample <- sample_n(df_cover_corrected, 4000)

base_plot <- ggplot(plot_sample,
                    aes(fill = cover, colour = cover),
                    alpha = 0.5
                    ) +
  theme(axis.text = element_text(size = 8), axis.title = element_text(size = 10),
        legend.text = element_text(size = 10)) +
  labs(fill = "", colour = "") +
  theme_bw() +
  guides(colour = guide_legend(override.aes = list(size = 2)))

pptMean_pptAnomaly <- base_plot + geom_point(aes(x = pptMean, y = pptAnomaly), size = 0.1) +
  labs(x = expression(pptMean ~ (mm ~ yr^-1)), y = "% annual anomaly") +
  ylim(-100, 150)
pptMean_int <- base_plot + geom_point(aes(x = pptMean, y = pptIntensity), size = 0.1) +
  labs(x = expression(pptMean ~ (mm ~ yr^-1)), y = expression(Intensity ~ (mm ~ d^-1)))
pptMean_season <- base_plot + geom_point(aes(x = pptMean, y = seasonLength), size = 0.1) +
  labs(x = expression(pptMean ~ (mm ~ yr^-1)), y = "Rainy season\nlength (d)")
pptMean_ugi <- base_plot + geom_point(aes(x = pptMean, y = pptUgi), size = 0.1) +
  labs(x = expression(pptMean ~ (mm ~ yr^-1)), y = "UGi")
pptMean_wet95 <- base_plot + geom_point(aes(x = pptMean, y = pptGt95Pc), size = 0.1) +
  labs(x = expression(pptMean ~ (mm ~ yr^-1)), y = expression(F["95w"]))

p1 <- ggarrange(pptMean_pptAnomaly, pptMean_ugi, pptMean_season, pptMean_int, pptMean_wet95,
          ncol = 2, nrow = 3, common.legend = TRUE)

gpp_expr <- expression(GPP ~ (gC ~ m^-2 ~ yr^-1))

gpp_sand <- base_plot + geom_point(aes(x = sand, y = gpp), size = 0.1, alpha = 0.5) +
  labs(x = "Soil sand fraction", y = gpp_expr)
gpp_clay <- base_plot + geom_point(aes(x = clay, y = gpp), size = 0.1, alpha = 0.5) +
  labs(x = "Soil clay fraction", y = gpp_expr)
gpp_T <- base_plot + geom_point(aes(x = tMean, y = gpp), size = 0.1, alpha = 0.5) +
  labs(x = "Mean air temperature\n(\u00B0C)", y = gpp_expr)
gpp_fire <- base_plot + geom_point(aes(x = fire, y = gpp), size = 0.1, alpha = 0.5) +
  labs(x = expression(Fire ~ frequency ~ (yr^-1)), y = gpp_expr)

gpp_pptMean <- base_plot + geom_point(aes(x = pptMean, y = gpp), size = 0.1, alpha = 0.5) +
  labs(x = expression(pptMean ~ (mm ~ yr^-1)), y = gpp_expr)
gpp_pptAnomaly <- base_plot + geom_point(aes(x = pptAnomaly, y = gpp), size = 0.1, alpha = 0.5) +
  labs(x = "% annual precipitation anomlay", y = gpp_expr) +
  xlim(-100, 150)
gpp_int <- base_plot + geom_point(aes(x = pptIntensity, y = gpp), size = 0.1, alpha = 0.5) +
  labs(x = expression(Precipitation ~ intensity ~(mm ~ d^-1)), y = gpp_expr)
gpp_dry <- base_plot + geom_point(aes(x = pptDryDays, y = gpp), size = 0.1, alpha = 0.5) +
  labs(x = "Fraction of dry\ndays per year", y = gpp_expr)
gpp_ugi <- base_plot + geom_point(aes(x = pptUgi, y = gpp), size = 0.1, alpha = 0.5) +
  labs(x = "UGi", y = gpp_expr)
gpp_wet95 <- base_plot + geom_point(aes(x = pptGt95Pc, y = gpp), size = 0.1, alpha = 0.5) +
  labs(x = expression(F["95w"]), y = gpp_expr)
gpp_season <- base_plot + geom_point(aes(x = seasonLength, y = gpp), size = 0.1, alpha = 0.5) +
  labs(x = "Wet season\nlength (days)", y = gpp_expr)
gpp_soilN <- base_plot + geom_point(aes(x = soilN, y = gpp), size = 0.1, alpha = 0.5) +
  labs(x = expression(Soil ~ nitrogen ~ (g ~ kg^-1)), y = gpp_expr)
gpp_soilP <- base_plot + geom_point(aes(x = soilP, y = gpp), size = 0.1, alpha = 0.5) +
  labs(x = "Soil extractable\nphosphorus (ppm)", y = gpp_expr)

p2 <- ggarrange(gpp_pptMean, gpp_pptAnomaly, gpp_int, gpp_wet95, gpp_ugi, gpp_season,
                nrow = 3, ncol = 2, common.legend = TRUE)

p3 <- ggarrange(gpp_T, gpp_soilN, gpp_sand, gpp_fire,
                nrow = 2, ncol = 2, common.legend = TRUE)

# Save figures to directory

ggsave(here(plot_dir, "p1.png"),
       plot = p1,
       width = 16, height = 20, units = "cm", dpi = 300)
ggsave(here(plot_dir, "p2.png"),
       plot = p2,
       width = 16, height = 20, units = "cm", dpi = 300)
ggsave(here(plot_dir, "p3.png"),
       plot = p3,
       width = 16, height = 18, units = "cm", dpi = 300)

```

Prepare raster layers for plotting - generate additional variables and take means
of annual variables.

```{r raster_prep}

# Define variable groups for summarising
static_var_names <- c("lc", "sand", "clay", "soilN", "soilP")

annual_var_names_for_mean <- c(
  "gpp",
  "pptTotal", "pptIntensity", "pptGt10", "pptGt30", "pptGt95Pc", "pptUgi", "seasonLength",
  "tMean", "modisBurn"
)

annual_var_names_for_sd <- c("pptTotal")

# Extract static variables
static_vars <- subset(var_stack_all, static_var_names)

# Extract variables where mean is taken across years
mean_vars <- map(annual_var_names_for_mean, function(var_name) {
  layer_names <- str_starts(names(var_stack_all), var_name)
  layers <- var_stack_all[[layer_names]]
  
  mean_layer <- mean(layers, na.rm = T)
  names(mean_layer) <- var_name
  
  mean_layer
}) %>% rast()

names(mean_vars)[2] <- "pptMean"

# Extract variables where SD is taken across years
sd_vars <- map(annual_var_names_for_sd, function(var_name) {
  layer_names <- str_starts(names(var_stack_all), var_name)
  layers <- var_stack_all[[layer_names]]
  
  sd_layer <- app(layers, sd, na.rm = T)
  names(sd_layer) <- var_name
  
  sd_layer
}) %>% rast()

names(sd_vars) <- c("pptSD")
sd_vars[["pptCV"]] <- sd_vars[["pptSD"]] / mean_vars[["pptMean"]]

# Merge back into single spatRaster object
var_stack <- c(static_vars, mean_vars, sd_vars)

# Mask to exclude "woody savannas"
var_stack_nonwoody <- mask(var_stack, var_stack$lc %in% c(7, 9, 10), maskvalues = c(0, NA))
values(study_area_mask)[values(var_stack$lc) == 8] <- 0

```


Exploratory maps of key covariates:


```{r eda_maps, eval = FALSE, echo = FALSE, results = "hide"}

plot_dir <- here("results", "figures")

# Create base map for all maps

africa_basemap <- tm_shape(africa) + tm_borders(col = "grey55") +
  tm_layout(main.title.position = 0.1,
            legend.position = c("left", "bottom"), legend.text.size = 0.5,
            legend.title.size = 0.85,
            legend.bg.color = "white", legend.bg.alpha = 0,
            legend.height = -0.4, legend.width = 0.6) +
  tm_graticules(x = c(-15, 0, 15, 30, 45), y = c(-30, -15, 0, 15, 30),
                col = "grey90", alpha = 0.6, labels.size = 0.6, lwd = 0.75)

# Create raster to mask excluded regions
outside_sa <- mask(study_area_mask, study_area_mask, maskvalues = c(1, NA))

sa_basemap <- tm_shape(outside_sa) + 
  tm_raster(palette = "grey85", legend.show = FALSE)

# Study area map
sa_map <- tm_shape(var_stack_nonwoody[["lc"]]) +
  tm_raster(style = "cat", palette = "-Set2",
            labels = as.character(igbp_index$cover), title = "Land cover class") +
  sa_basemap +
  africa_basemap +
  tm_layout(main.title = "d)")

# GPP across region
gpp_map <- tm_shape(var_stack_nonwoody[["gpp"]]) + 
  tm_raster(n = 9, palette = "YlGn", style = "cont", title = expression(GPP ~ (gC ~ m^-2 ~ yr^-1)),
            breaks = seq(0, 2500, 500)) +
  sa_basemap +
  africa_basemap +
  tm_layout(main.title = "a)")

# Precipitation maps

map_map <- tm_shape(var_stack_nonwoody[["pptMean"]]) +
  tm_raster(style = "cont", palette = "YlGnBu", title = expression("MAP" ~ (mm ~ yr^-1)),
            breaks = seq(0, 1500, 300)) +
  sa_basemap +
  africa_basemap +
  tm_layout(main.title = "b)")

cv_map <- tm_shape(var_stack_nonwoody[["pptCV"]]) +
  tm_raster(style = "cont", n = 10, palette = "RdPu", title = "CV of annual\nprecipitation (%)",
            breaks = seq(0, 0.60, 0.10)) +
  sa_basemap +
  africa_basemap

intensity_map <- tm_shape(var_stack_nonwoody[["pptIntensity"]]) +
  tm_raster(style = "cont", palette = "PRGn", title = expression(Intensity ~ (mm ~ day^-1)),
            breaks = seq(0, 25, 5)) +
  sa_basemap +
  africa_basemap

ugi_map <- tm_shape(var_stack_nonwoody[["pptUgi"]]) +
  tm_raster(style = "cont", n = 10, palette = "-PRGn", title = "UGi of daily\nprecipitation",
            breaks = seq(0.1, 0.6, 0.1), midpoint = 0.35) +
  sa_basemap +
  africa_basemap

wet95_map <- tm_shape(var_stack_nonwoody[["pptGt95Pc"]]) +
  tm_raster(style = "cont", n = 10, palette = "OrRd", title = "Fraction of precipitation\nover 95th percentile",
            breaks = seq(0, 0.25, 0.05)) +
  sa_basemap +
  africa_basemap

# Seasonality maps

season_map <- tm_shape(var_stack_nonwoody[["seasonLength"]] %>% mask(var_stack_nonwoody[["seasonLength"]] > 0, maskvalues = c(NA, 0))) +
  tm_raster(style = "cont", palette = "RdYlBu", title = "Rainy season\nlength (days)",
            breaks = seq(0, 300, 50), midpoint = 150) +
  sa_basemap +
  africa_basemap

# Soil property maps

soilN_map <- tm_shape(var_stack_nonwoody[["soilN"]]) +
  tm_raster(style = "cont", palette = "YlGn", title = expression(soilN ~ (g ~ kg^-1)),
            breaks = seq(0, 2, 0.4)) +
  sa_basemap +
  africa_basemap

soilP_map <- tm_shape(var_stack_nonwoody[["soilP"]]) +
  tm_raster(style = "cont", palette = "YlGn", title = expression(soilP ~ (g ~ kg^-1)),
            breaks = seq(0, 25, 5)) +
  sa_basemap +
  africa_basemap

sand_map <- tm_shape(var_stack_nonwoody[["sand"]]) +
  tm_raster(style = "cont", palette = "-BrBG", title = "Soil sand %",
            breaks = seq(0, 100, 20)) +
  sa_basemap +
  africa_basemap

clay_map <- tm_shape(var_stack_nonwoody[["clay"]]) +
  tm_raster(style = "cont", palette = "-BrBG", title = "Soil clay %",
            breaks = seq(0, 100, 20)) +
  sa_basemap +
  africa_basemap

# Temperature maps

T_map <- tm_shape(var_stack_nonwoody[["tMean"]]) +
  tm_raster(style = "cont", n = 10, palette = "-RdYlBu", title = "Mean air\ntemperature (\u00B0C)",
            breaks = seq(10, 35, 5)) +
  sa_basemap +
  africa_basemap +
  tm_layout(main.title = "c)")

# Fire frequency map

fire_map <- tm_shape(var_stack_nonwoody[["modisBurn"]]) +
  tm_raster(style = "cont", palette = "YlOrRd", title = expression(Fire ~ frequency ~ (yr^-1)),
            breaks = seq(0, 0.8, 0.2)) +
  sa_basemap +
  africa_basemap

# Save map figures to directory

tmap_save(season_map,
          here(plot_dir, "season_map_plot.png"),
          width = 16, height = 12, units = "cm", dpi = 300)

multi_plot <- tmap_arrange(gpp_map, map_map, T_map, sa_map)

tmap_save(multi_plot,
          here(plot_dir, "multi_map_plot.png"),
          width = 18, height = 18, units = "cm", dpi = 300)

```
