==> analysis/R/4_africa_rangelands_gam_analysis_annual.Rmd <==
---
title: Disentangling the environmental drivers of gross primary productivity in African rangelands - annual analysis
author: Guy Lomax
date: 2024-05-14
output: html_notebook
---

This analysis uses Generalised Additive Models (GAMs) to explore the
relationship between different dimensions of precipitation variability and
vegetation productivity across African rangelands. This notebook models annually
integrated gross primary productivity each year from 2001 to 2019 as a function of
precipitation, temperature and other metrics for each rangeland pixel in Sub-
Saharan Africa.

GPP data are from PML_V2 dataset (Gan et al., 2018; Zhang et al., 2019) at 8-day
temporal and 500m spatial resolution.
Precipitation data are from CHIRPS (Funk et al., 2015) at daily temporal
resolution and 0.05 degree spatial resolution.
All analysis conducted at 0.05 degree spatial resolution. 

```{r setup, include=FALSE}

# If running for first time, run renv::restore() to install required packages
# renv::restore()

# Data management

library(tidyverse)
library(terra)
library(sf)
library(here)

# Analysis
library(parallel)
library(mgcv)
library(tictoc)

# Visualisation
library(tmap)
library(gratia)
library(ggpubr)
library(patchwork)

# Directory to save figures
plot_dir <- here("results", "figures")

ggplot2::theme_set(theme_bw())

# Establish number of cores for parallel computing
nc <- detectCores() / 4

```

# Load and pre-process data

Raster data layers prepared, cropped and aggregated to 0.05 deg resolution
using Google Earth Engine.


```{r load_data, results = "hide", include = FALSE}

# Load multi-annual data frame for modelling

df <- read_csv(here("data", "processed", "csv", "df_annual.csv")) %>%
  select(cell, x, y, gpp, starts_with("ppt"), starts_with("season"), everything())


```

# Exploratory Data Analysis

Plot histograms and scatter plots of key covariates.

Variable definitions:

|Variable name|Definition|Unit|Source dataset|
|-------------|----------|----|--------------|
|gpp|Gross Primary Productivity|gC m-2 yr-1|PML_V2|
|-------------|----------|----|--------------|
|pptMean|Mean annual precipitation|mm yr-1|CHIRPS|
|pptAnomaly|Ratio of annual to mean precipitation|unitless|CHIRPS|
|pptDryDays|Fraction of days per year with < 1mm precipitation|unitless 0-1|CHIRPS|
|pptIntensity|Mean intensity of rainfall on a rainy day|mm day-1|CHIRPS|
|pptGt95Pc|Fraction of daily rainfall above 95th percentile of wet days|unitless 0-1|CHIRPS|
|pptGt10|Fraction of days per year with >10 mm rainfall|
|pptGt30|Fraction of days per year with >30 mm rainfall|
|pptUgi|Unranked Gini index of daily precipitation|unitless 0-1|CHIRPS|
|seasonLength|Average rainy season length|days|CHIRPS|
|-------------|----------|----|
|sand|Soil sand fraction 0-20cm|%|iSDAsoil|
|clay|Soil clay fraction 0-20cm|%|iSDAsoil|
|soilN|Soil nitrogen content 0-20cm|g/kg|iSDAsoil|
|soilP|Soil extractable phosphorus 0-20cm|ppm|iSDAsoil|
|tMean|Mean annual air temperature at 2m|degC|ERA5-Land|
|fire|Mean fire frequency 2001-2019|yr-1|MODIS Terra|


Calculate correlation matrix for all variables in dataset:

```{r correlation, eval = FALSE}

ppt_var_names <- c("pptMean", "pptAnomaly", "pptIntensity", "pptDryDays", "pptUgi", 
                   "pptGt95Pc", "pptGt10", "pptGt30", "seasonLength")
non_ppt_var_names <- c("sand", "clay", "soilN", "soilP", "tMean", "fire")

cor_matrix <- df %>%
  dplyr::select(gpp, all_of(ppt_var_names), all_of(non_ppt_var_names)) %>%
  cor(method = "spearman", use="pairwise.complete.obs")

var_names <- c("GPP", "MAP", "anomaly", "intensity", "dry", "UGi",
               ":F['95w']", ":D['10mm']", ":D['30mm']", "season",
               "sand", "clay", "soilN", "soilP", "tMean", "fire")
colnames(cor_matrix) <- var_names
rownames(cor_matrix) <- var_names

# Save correlation plot figure to directory

png(paste0(plot_dir, "/cor_plot_annual.png"),
     height = 50, width = 50, units = "cm", res = 300)

corrplot::corrplot(cor_matrix, method = "number", type = "upper",
                               tl.cex = 3, number.cex = 1.8, cl.cex = 2)

while (!is.null(dev.list()))  dev.off()


```

We remove covariates with > 0.7 Spearman's Rank correlation coefficient with
other covariates, leaving the following:

- Mean annual precipitation
- Precipitation anomaly
- Precipitaton mean day anomaly
- Precipitation intensity
- F95w
- UGi
- Season length
- Mean air temperature
- Soil sand fraction
- Soil N content
- Fire occurrence

# GAMs

We use GAMs to explore the relationships and interactions between
rainfall dimensions, non-rainfall covariates and GPP.

The dependent variable (GPP) is non-negative and variance tends to increase with
the mean. A Gamma distribution with a log link may therefore be appropriate
for modelling these data.

We also include a smooth function of latitude and longitude te(x,y) to include
spatial relationships and spatially correlated errors not accounted for by other
covariates, as well as a smooth function of time s(year)

To prepare for GAM modelling, we select only covariates retained from correlation
analysis above. We also take a random sample of 10% of the dataset to reduce
model fitting issues.

```{r gam_prep}

# Select variables equivalent to multi-annual analysis

target_vars <- c(
  "cover",
  "gpp",
  "pptTotal", "pptMean", "pptAnomaly",
  "pptIntensity", "pptGt95Pc", "pptUgi",
  "seasonLength",
  "tMean", "fire", "sand", "soilN"
)

df_model <- df %>%
  select(cell, x, y, year, all_of(target_vars)) %>%
  filter(cover != "Woody savannas") %>%
  mutate(pptTotal = 0.001 * pptTotal,
         pptMean = 0.001 * pptMean,
         cover = str_replace(cover, " ", "_"))

# Subset of 10% of rows to reduce model fitting issues
set.seed(123)

df_model_sample <- sample_frac(df_model, 0.1)

write_csv(df_model,
          here("data", "processed", "csv", "df_model_annual.csv"))
write_csv(df_model_sample,
          here("data", "processed", "csv", "df_model_annual_sample.csv"))


```


Fit four models for each land cover type: one without a spatial (x,y) term, and
three with different degrees of complexity allowed (k = c(8,8) to k = c(16,16)).
The range of results of these models accounts for spatial autocorrelation and
allows a robustness check of the results to the degree of spatial structure.

```{r standard_models, eval = FALSE}

df_model_sample <- read_csv(here("data", "processed", "csv", "df_model_annual_sample.csv")) 

cl <- makeCluster(nc)

# Loop through land cover classes to fit models

cover_classes <- unique(df_model_sample$cover)

for (i in cover_classes) {
  
  df_model_class <- filter(df_model_sample, cover == i)
  
  message("Fitting model 1 for ", i)
  tic()
  mod1 <- bam(gpp ~ s(pptMean, k = 12, bs = "ts") +
              s(pptAnomaly, k = 8, bs = "ts") +
              s(pptIntensity, k = 8, bs = "ts") +
              s(pptGt95Pc, k = 8, bs = "ts") +
              s(pptUgi, k = 8, bs = "ts") +
              s(seasonLength, k = 8, bs = "ts") +
              s(tMean, k = 8, bs = "ts") +
              s(fire, k = 8, bs = "ts") +
              s(sand, k = 8, bs = "ts") +
              s(soilN, k = 8, bs = "ts") +
              s(year, k = 8, bs = "ts"),
            data = df_model_class,
            method = "REML",
            family = Gamma(link = "log"),
            select = TRUE,
            samfrac = 0.1,
            cluster = cl,
            nthreads = 4)
  toc()

  write_rds(mod1, here("data", "processed", "models", paste0("mod1_", i, "_annual.rds")))

  message("Fitting model 2k8 for ", i)
  tic()
  mod2_k8 <- bam(gpp ~ s(pptMean, k = 12, bs = "ts") +
              s(pptAnomaly, k = 8, bs = "ts") +
              s(pptIntensity, k = 8, bs = "ts") +
              s(pptGt95Pc, k = 8, bs = "ts") +
              s(pptUgi, k = 8, bs = "ts") +
              s(seasonLength, k = 8, bs = "ts") +
              s(tMean, k = 8, bs = "ts") +
              s(fire, k = 8, bs = "ts") +
              s(sand, k = 8, bs = "ts") +
              s(soilN, k = 8, bs = "ts") +
              te(x, y, k = c(8, 8), bs = "ts") +
              s(year, k = 8, bs = "ts"),
            data = df_model_class,
            method = "fREML",
            family = Gamma(link = "log"),
            select = TRUE,
            samfrac = 0.1,
            cluster = cl,
            nthreads = 4)
  toc()

  write_rds(mod2_k8, here("data", "processed", "models", paste0("mod2_k8_", i, "_annual.rds")))

  message("Fitting model 2k12 for ", i)
  tic()
  mod2_k12 <- bam(gpp ~ s(pptMean, k = 12, bs = "ts") +
              s(pptAnomaly, k = 8, bs = "ts") +
              s(pptIntensity, k = 8, bs = "ts") +
              s(pptGt95Pc, k = 8, bs = "ts") +
              s(pptUgi, k = 8, bs = "ts") +
              s(seasonLength, k = 8, bs = "ts") +
              s(tMean, k = 8, bs = "ts") +
              s(fire, k = 8, bs = "ts") +
              s(sand, k = 8, bs = "ts") +
              s(soilN, k = 8, bs = "ts") +
              te(x, y, k = c(12, 12), bs = "ts") +
              s(year, k = 8, bs = "ts"),
            data = df_model_class,
            method = "fREML",
            family = Gamma(link = "log"),
            select = TRUE,
            samfrac = 0.1,
            cluster = cl,
            nthreads = 4)
  toc()

  write_rds(mod2_k12, here("data", "processed", "models", paste0("mod2_k12_", i, "_annual.rds")))

  message("Fitting model 2k16 for ", i)
  tic()
  mod2_k16 <- bam(gpp ~ s(pptMean, k = 12, bs = "ts") +
              s(pptAnomaly, k = 8, bs = "ts") +
              s(pptIntensity, k = 8, bs = "ts") +
              s(pptGt95Pc, k = 8, bs = "ts") +
              s(pptUgi, k = 8, bs = "ts") +
              s(seasonLength, k = 8, bs = "ts") +
              s(tMean, k = 8, bs = "ts") +
              s(fire, k = 8, bs = "ts") +
              s(sand, k = 8, bs = "ts") +
              s(soilN, k = 8, bs = "ts") +
              te(x, y, k = c(16, 16), bs = "ts") +
              s(year, k = 8, bs = "ts"),
            data = df_model_class,
            method = "fREML",
            family = Gamma(link = "log"),
            select = TRUE,
            samfrac = 0.1,
            cluster = cl,
            nthreads = 4)
  toc()

  write_rds(mod2_k16, here("data", "processed", "models", paste0("mod2_k16_", i, "_annual.rds")))

}

stopCluster(cl)

```

Load saved models and run model diagnostics

```{r standard_models_load}

mod1_grass <- read_rds(here("data", "processed", "models", "mod1_Grasslands_annual.rds"))
mod2_k8_grass <- read_rds(here("data", "processed", "models", "mod2_k8_Grasslands_annual.rds"))
mod2_k12_grass <- read_rds(here("data", "processed", "models", "mod2_k12_Grasslands_annual.rds"))
mod2_k16_grass <- read_rds(here("data", "processed", "models", "mod2_k16_Grasslands_annual.rds"))

mod1_shrub <- read_rds(here("data", "processed", "models", "mod1_Open_shrublands_annual.rds"))
mod2_k8_shrub <- read_rds(here("data", "processed", "models", "mod2_k8_Open_shrublands_annual.rds"))
mod2_k12_shrub <- read_rds(here("data", "processed", "models", "mod2_k12_Open_shrublands_annual.rds"))
mod2_k16_shrub <- read_rds(here("data", "processed", "models", "mod2_k16_Open_shrublands_annual.rds"))

mod1_savanna <- read_rds(here("data", "processed", "models", "mod1_Open_savannas_annual.rds"))
mod2_k8_savanna <- read_rds(here("data", "processed", "models", "mod2_k8_Open_savannas_annual.rds"))
mod2_k12_savanna <- read_rds(here("data", "processed", "models", "mod2_k12_Open_savannas_annual.rds"))
mod2_k16_savanna <- read_rds(here("data", "processed", "models", "mod2_k16_Open_savannas_annual.rds"))

```



# Plot GAM results

```{r term_plots}

target_vars <- c("pptMean", "pptAnomaly", "pptIntensity", "tMean", "soilN", "sand")

df_model_sample <- read_csv(here("data", "processed", "csv", "df_model_annual_sample.csv"))

cover_list <- c("Shrublands", "Savannas", "Grasslands")
complexity_list <- c(0, 8, 12, 16)

df_model_sample$cover[df_model_sample$cover == "Open_shrublands"] <- "Shrublands"
df_model_sample$cover[df_model_sample$cover == "Open_savannas"] <- "Savannas"
df_model_sample_clean <- df_model_sample %>%
  filter(cover != "Woody_savannas") %>%
  mutate(cover = ordered(cover, levels = cover_list)) %>%
  mutate(pptMean = pptMean * 1000) %>%
  mutate(pptAnomaly = (pptAnomaly - 1) * 100)

# Identify 0.5th and 99.5th percentile of data in each class - exclude from
# plots as per R2 suggestion

df_model_bounds <- df_model_sample_clean %>%
  select(cover, all_of(target_vars)) %>%
  pivot_longer(all_of(target_vars), names_to = "variable") %>%
  group_by(cover, variable) %>%
  summarise(low_bound = quantile(value, 0.005),
            high_bound = quantile(value, 0.995))

# Set plot output directory

plot_dir <- here("results", "figures")

theme_set(theme_classic())

# Variable names and labels

var_labels <- read_csv(here("data", "raw", "csv", "var_labels_annual.csv"))

var_list <- var_labels$vars

## Main GAM partial effect plots

# Extract smooth shapes as data frame using gratia

mod_list <- list(mod1_shrub, mod2_k8_shrub, mod2_k12_shrub, mod2_k16_shrub,
                 mod1_savanna, mod2_k8_savanna, mod2_k12_savanna, mod2_k16_savanna,
                 mod1_grass, mod2_k8_grass, mod2_k12_grass, mod2_k16_grass)

mod_smooths_all <- tibble(cover = rep(cover_list, each = 4),
                          complexity = rep(complexity_list, 3),
                          models = mod_list) %>%
  mutate(smooths = map(models, gratia::smooth_estimates, unconditional = TRUE)) %>%
  select(-models) %>%
  unnest(smooths) %>%
  mutate(ci_low = est - se * 1.96,
         ci_high = est + se * 1.96,
         pptMean = pptMean * 1000,
         pptAnomaly = (pptAnomaly - 1) * 100)

mod_smooths_summary <- mod_smooths_all %>%
  select(-by, -type, -smooth, -x, -y, -year) %>%
  pivot_longer(cols = all_of(var_list), names_to = "variable", values_to = "value") %>%
  drop_na() %>%
  filter(complexity != 0) %>%  # Exclude non-spatial models from plots
  group_by(cover, variable, value) %>%
  summarise(est = exp(mean(est)),
            ci_low = exp(min(ci_low)),
            ci_high = exp(max(ci_high))) %>%
  ungroup()

mod_smooths_summary_bounds <- mod_smooths_summary %>%
  left_join(df_model_bounds) %>%
  filter(value >= low_bound & value <= high_bound)

palette <- "Set2"

# Generate plots

plots <- map(target_vars, function(var) {
  
  # Labels for each facet window
  # Create expressions for those with superscript/subscripts
  x_label <- var_labels$label[var_labels$vars == var]
  x_label <- parse(text = x_label)
  
  # Plot data density plot
  data_df <- df_model_sample_clean %>%
    select(x, y, cover, all_of(var)) %>%
    rename(z = 4) %>%
    left_join(df_model_bounds %>% filter(variable == var)) %>%
    filter(z >= low_bound & z <= high_bound)
  
  density <- ggplot(data_df, aes(x = z)) +
    geom_density(aes(colour = cover, fill = cover),
                 position = "identity",
                 alpha = 0.2, 
                 linewidth = 0.75,
                 show.legend = FALSE) +
    scale_colour_brewer(palette = palette, direction = -1) +
    scale_fill_brewer(palette = palette, direction = -1) +
    theme_void()
  
  # Plot model smooths
  smooth_df <- mod_smooths_summary_bounds %>%
    filter(variable == var) %>%
    mutate(cover = ordered(cover, levels = cover_list))
  
  plot <- ggplot(smooth_df, aes(x = value, y = est)) +
    geom_hline(yintercept = 1, colour = "grey80", linetype = "longdash") +
    geom_ribbon(aes(ymin = ci_low, ymax = ci_high, fill = cover), alpha = 0.3) +
    geom_line(aes(colour = cover), lwd = 0.75) +
    scale_colour_brewer(palette = palette, direction = -1) +
    scale_fill_brewer(palette = palette, direction = -1) +
    coord_cartesian(ylim=c(0, 2.5)) +
    labs(x = x_label,
         y = "Partial effect",
         colour = "Land cover",
         fill = "Land cover"
         )
  
  if(var_labels$min_0[var_labels$vars == var] == 1) {
    plot <- plot + scale_x_continuous(limits = c(0, max(smooth_df$value)))
  }
  
  if(var_labels$midpoint[var_labels$vars == var] != -1) {
    midpoint <- var_labels$midpoint[var_labels$vars == var]
    plot <- plot + geom_vline(xintercept = midpoint, colour = "grey80", linetype = "longdash")
  }
  
  # Arrange into small multiples with patchwork package 
  density + plot +
    plot_layout(ncol = 1, nrow = 2, widths = 4, heights = c(1.5, 4))
})

# Create main GAM figure
gam_plot <- ggarrange(plotlist = plots, ncol = 3, nrow = 2, common.legend = TRUE)

# Save plot
ggsave(here(plot_dir, "gam_plot_signif_sp_annual.jpg"),
       plot = gam_plot,
       width = 20, height = 30, units = "cm", dpi = 300)


```



Model diagnostic and residual plots

```{r diag_plots, eval = FALSE}

# Diagnostic plots
mod_diag_grass <- appraise(mod2_k12_grass, ci_alpha = 0.5, point_alpha = 0.05, n_bins = 50)
mod_diag_shrub <- appraise(mod2_k12_shrub, ci_alpha = 0.5, point_alpha = 0.05, n_bins = 50)
mod_diag_savanna <- appraise(mod2_k12_savanna, ci_alpha = 0.5, point_alpha = 0.05, n_bins = 50)


ggsave(here("results", "figures", "mod1_diag_grass.png"), mod_diag_grass,
       width = 26.5, height = 22, units = "cm", dpi = 250)
ggsave(here("results", "figures", "mod_diag_shrub.png"), mod_diag_shrub,
       width = 26.5, height = 22, units = "cm", dpi = 250)
ggsave(here("results", "figures", "mod_diag_savanna.png"), mod_diag_savanna,
       width = 26.5, height = 22, units = "cm", dpi = 250)

# Concurvity plots
cc_shrub <- concurvity(mod1_shrub, full = FALSE)
cc_savanna <- concurvity(mod1_savanna, full = FALSE)
cc_grass <- concurvity(mod1_grass, full = FALSE)

save_corplot <- function(matrix, filename) {
  png(paste0("results/figures/", filename),
     height = 50, width = 50, units = "cm", res = 300)
  
  corrplot::corrplot(matrix, method = "number", type = "upper",
                     tl.cex = 3, number.cex = 1.8, cl.cex = 2,
                     is.corr = FALSE,
                     col = viridis::magma(n = 20, begin = 0.3, end = 0.8)
                     )
  
  while (!is.null(dev.list()))  dev.off()
  
}

# Save plots (may need to run in terminal due to issue with running in R Markdown)
save_corplot(cc_shrub$estimate, "cc_shrub.png")
save_corplot(cc_savanna$estimate, "cc_savanna.png")
save_corplot(cc_grass$estimate, "cc_grass.png")

```



Repeat model for separate MAP bins:

```{r binned_models, eval = FALSE}

df_model_sample <- read_csv(here("data", "processed", "csv", "df_model_annual_sample.csv"))

df_model_bins <- df_model_sample %>%
  filter(cover != "Woody_savannas") %>%
  mutate(map_bin = cut_width(pptMean, width = 0.4, boundary = 0)) %>%
  group_by(cover, map_bin) %>%
  nest() %>%
  mutate(frac = map(data, nrow) %>% unlist() / nrow(df_model_sample)) %>%
  filter(frac > 0.01) %>%
  unnest(data)

k_list <- c(8, 12, 16)

fit_gam <- function(data, spatial = FALSE, spatial_k = 12) {
  
  if (spatial) {
    model <- bam(gpp ~ s(pptMean, k = 12, bs = "ts") +
                   s(pptAnomaly, k = 8, bs = "ts") +
                   s(pptIntensity, k = 8, bs = "ts") +
                   s(pptGt95Pc, k = 8, bs = "ts") +
                   s(pptUgi, k = 8, bs = "ts") +
                   s(seasonLength, k = 8, bs = "ts") +
                   s(tMean, k = 8, bs = "ts") +
                   s(fire, k = 8, bs = "ts") +
                   s(sand, k = 8, bs = "ts") +
                   s(soilN, k = 8, bs = "ts") +
                   te(x, y, k = c(spatial_k, spatial_k), bs = "ts") +
                   s(year, k = 8, bs = "ts"),
                 data = data,
                 method = "fREML",
                 family = Gamma(link = "log"),
                 select = TRUE,
                 samfrac = 0.1,
                 cluster = cl,
                 nthreads = 4)
  } else {
    model <- bam(gpp ~ s(pptMean, k = 12, bs = "ts") +
                   s(pptAnomaly, k = 8, bs = "ts") +
                   s(pptIntensity, k = 8, bs = "ts") +
                   s(pptGt95Pc, k = 8, bs = "ts") +
                   s(pptUgi, k = 8, bs = "ts") +
                   s(seasonLength, k = 8, bs = "ts") +
                   s(tMean, k = 8, bs = "ts") +
                   s(fire, k = 8, bs = "ts") +
                   s(sand, k = 8, bs = "ts") +
                   s(soilN, k = 8, bs = "ts") +
                   s(year, k = 8, bs = "ts"),
                 data = data,
                 method = "fREML",
                 family = Gamma(link = "log"),
                 select = TRUE,
                 samfrac = 0.1,
                 cluster = cl,
                 nthreads = 4)
  }
  
  model
}

cl <- makeCluster(nc)

mod1_binned <- df_model_bins %>%
  group_by(cover, map_bin) %>%
  nest() %>%
  mutate(mod = map(data, fit_gam, spatial = FALSE)) %>%
  select(-data)

write_rds(mod1_binned, here("data", "processed", "models", "mod_binned_nsp_lc_subsets.rds"))

for (i in k_list) {

  colname <- paste0("mod_k", i)

  message("Fitting models for k = ", i)

  tic()
  mod_binned <- df_model_bins %>%
    group_by(cover, map_bin) %>%
    nest() %>%
    mutate(mod = map(data, fit_gam, spatial = TRUE, spatial_k = i)) %>%
    select(-data) %>%
    rename_with(.cols = "mod", .fn = ~paste0(., "_", i))

  write_rds(mod_binned, here("data", "processed", "models", paste0("mod_binned_k", i, "_lc_subsets.rds")))
  toc()
}

stopCluster(cl)

```

Visualise binned models:

```{r binned_models_viz}

target_vars <- c("pptAnomaly", "pptIntensity", "tMean", "soilN", "sand")

# Load models and generate smooths
# (Done in one step to avoid issues with keeping large models in memory)

mod1_smooths_binned <- read_rds(here("data", "processed", "models", "mod_binned_nsp_lc_subsets.rds")) %>%
  filter(map_bin %in% levels(map_bin)[1:3]) %>%
  mutate(smooths_df = map(mod, smooth_estimates, unconditional = TRUE, n = 100)) %>%
  select(-mod) %>%
  unnest(smooths_df) %>%
  mutate(ci_low = est - se * 1.96,
         ci_high = est + se * 1.96,
         pptMean = pptMean * 1000,
         pptAnomaly = (pptAnomaly - 1) * 100,
         k = 0)

mod_smooths_binned_k8 <- read_rds(here("data", "processed", "models", "mod_binned_k8_lc_subsets.rds")) %>%
  filter(map_bin %in% levels(map_bin)[1:3]) %>%
  mutate(smooths_df = map(mod_8, smooth_estimates, unconditional = TRUE, n = 100)) %>%
  select(-mod_8) %>%
  unnest(smooths_df) %>%
  mutate(ci_low = est - se * 1.96,
         ci_high = est + se * 1.96,
         pptMean = pptMean * 1000,
         pptAnomaly = (pptAnomaly - 1) * 100,
         k = 8)

mod_smooths_binned_k12 <- read_rds(here("data", "processed", "models", "mod_binned_k12_lc_subsets.rds")) %>%
  filter(map_bin %in% levels(map_bin)[1:3]) %>%
  mutate(smooths_df = map(mod_12, smooth_estimates, unconditional = TRUE, n = 100)) %>%
  select(-mod_12) %>%
  unnest(smooths_df) %>%
  mutate(ci_low = est - se * 1.96,
         ci_high = est + se * 1.96,
         pptMean = pptMean * 1000,
         pptAnomaly = (pptAnomaly - 1) * 100,
         k = 12)

mod_smooths_binned_k16 <- read_rds(here("data", "processed", "models", "mod_binned_k16_lc_subsets.rds")) %>%
  filter(map_bin %in% levels(map_bin)[1:3]) %>%
  mutate(smooths_df = map(mod_16, smooth_estimates, unconditional = TRUE, n = 100)) %>%
  select(-mod_16) %>%
  unnest(smooths_df) %>%
  mutate(ci_low = est - se * 1.96,
         ci_high = est + se * 1.96,
         pptMean = pptMean * 1000,
         pptAnomaly = (pptAnomaly - 1) * 100,
         k = 16)

# Merge

mod_smooths_binned_all <- bind_rows(mod1_smooths_binned, mod_smooths_binned_k8, mod_smooths_binned_k12, mod_smooths_binned_k16)

mod_smooths_binned_all$cover[mod_smooths_binned_all$cover == "Open_shrublands"] <- "Shrublands"
mod_smooths_binned_all$cover[mod_smooths_binned_all$cover == "Open_savannas"] <- "Savannas"

# Identify 0.5th and 99.5th percentile of data in each class - exclude from
# plots as per R2 suggestion

df_binned_bounds <- df_model_sample_clean %>%
  select(cover, pptMean, all_of(target_vars)) %>%
  pivot_longer(all_of(target_vars), names_to = "var") %>%
  mutate(map_bin = cut_width(pptMean / 1000, 0.4, boundary = 0)) %>%
  filter(map_bin %in% levels(map_bin)[1:3]) %>%
  group_by(map_bin, cover, var) %>%
  summarise(low_bound = quantile(value, 0.005),
            high_bound = quantile(value, 0.995))

# Set plot output directory

plot_dir <- here("results", "figures")

theme_set(theme_classic())

## Binned GAM partial effect plots

smooths_binned_pivot <- mod_smooths_binned_all %>%
  select(cover, map_bin, k,
         all_of(target_vars),
         est, se, ci_low, ci_high) %>%
  pivot_longer(cols = all_of(target_vars),
               names_to = "var", values_to = "value") %>%
  drop_na()

smooths_binned_merged <- smooths_binned_pivot %>%
  filter(k != 0) %>%
  group_by(cover, map_bin, var, value) %>%
  summarise(est = exp(mean(est)),
            ci_low = exp(min(ci_low)),
            ci_high = exp(max(ci_high))) %>%
  ungroup()

# Label lookup tables

var_labels <- read_csv(here("data", "raw", "csv", "var_labels_annual.csv"))

target_var_labels <- var_labels %>%
  filter(vars %in% target_vars)

bin_labels <- c("'0-400'~mm~yr^-1",
                "'400-800'~mm~yr^-1",
                "'800-1,200'~mm~yr^-1")

var_labels_df <- tibble(
  var = target_vars,
  var_label = ordered(target_var_labels$label, levels = target_var_labels$label)
)

bin_labels_df <- tibble(
  map_bin = levels(smooths_binned_pivot$map_bin)[1:3],
  bin_label = ordered(bin_labels, levels = bin_labels)
)

smooths_binned_labeled <- smooths_binned_merged %>%
  left_join(var_labels_df) %>%
  left_join(bin_labels_df) %>%
  mutate(cover = ordered(cover, levels = c("Shrublands", "Savannas", "Grasslands")))

smooths_binned_bounds <- smooths_binned_labeled %>%
  left_join(df_binned_bounds) %>%
  filter(value >= low_bound & value <= high_bound)

# Generate faceted plot
palette <- "Set2"

plots_binned <- ggplot(smooths_binned_bounds, aes(x = value)) +
  geom_hline(yintercept = 1, colour = "grey80", linetype = "longdash") +
  geom_line(aes(y = est, colour = cover), lwd = 0.75) +
  geom_ribbon(aes(ymin = ci_low, ymax = ci_high, fill = cover), alpha = 0.3) +
  scale_colour_brewer(palette = palette, direction = -1) +
  scale_fill_brewer(palette = palette, direction = -1) +
  coord_cartesian(ylim = c(0.5, 1.5)) +
  labs(x = "Covariate value",
       y = "Partial effect",
       colour = "Land cover",
       fill = "Land cover") +
  facet_grid(rows = vars(bin_label), cols = vars(var_label),
             scales = "free_x",
             labeller = label_parsed,
             switch = "y") +
  theme_classic() +
  theme(panel.background = element_rect(fill = NA, color = "black"),
        legend.position = "top")

# Save to disk
ggsave(here("results", "figures", "plot_binned_signif_sp_annual.jpg"),
       plots_binned,
       width = 24, height = 16, units = "cm", dpi = 300)

```
